+++ routes/api/contacts.js
@@ -1,81 +1,18 @@
 const express = require('express');
 const router = express.Router();
-const Joi = require('joi');
 
-const { HttpError } = require('../../helpers')
-const contact = require('../../models/contacts')
+const controllers = require('../../controllers/contacts');
+const {validateBody} = require('../../middlewares');
+const schema = require('../../schemas/contactSchema');
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-})
+router.get('/', controllers.getAll);
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contact.listContacts();
-    res.json(result)
-  } catch (error) {
-    next(error)
-  }
-})
+router.get('/:id', controllers.getById);
 
-router.get('/:id', async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contact.getContactById(id);
-    if (!result) {
-      throw HttpError(404, "Not found")
-    }
-    res.json(result)
-  } catch (error) {
-    next(error)
-  }
-})
+router.post('/', validateBody(schema), controllers.addContact);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body)
-    if (error) {
-      throw HttpError(400, "missing required name field")
-    }
-    const result = await contact.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-})
+router.delete('/:id', controllers.removeContact);
 
-router.delete('/:id', async (req, res, next) => {
-  try {
-    const {id} = req.params;
-    const result = contact.removeContact(id);
-    if (!result) {
-      throw HttpError(404, "Not found")
-    }
-    res.json({
-      message: "contact deleted"
-    })
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.put('/:id', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body)
-    if (error) {
-      throw HttpError(400, error.message)
-    }
-    const {id} = req.params;
-    const result = await contact.updateContact(id, req.body);
-    if (!result) {
-      throw HttpError(404, "Not found")
-    }
-    res.json(result)
-  } catch (error) {
-    next(error);
-  }
-})
+router.put('/:id',  validateBody(schema), controllers.updateContact);
 
 module.exports = router
